Tab order editor / button to insert code
Changing the tab order in the jot editor So far, the permission is next to the message tab. However, it is important for the user to first check what the message looks like after creating a post. Images are added and the permission is only checked at the end. 

In addition, a button has been added that allows the BBCode for code to be inserted.