Update browser.js
Fix for Vier and other themes that use default Photo Browser modal and cannot navigate to albums.

`location.reload()` appeared to be overriding the `location.href` just above it that actually take you to the album. So it was taking you there but then immediately reloading the modal back to the "Recent Photos" page.