add contextmenu action to compose-button
when visiting a contact i can create a posting, mentioning this user.
Many times i visited a user and wanted to create a "normal" posting.
So i had to go back and push the create button.

This commit adds on desktop-view a right-click event and on
smartphone-view a longtap event (contextMenu) which opens the jot with
NO mention to create a "normal" posting from contact vcard.