fix file browser icons
In the file browser when viewing attachments the icons are missing. This was due to a misplaced parenthesis and "icon" instead of "icons" (plural) in the image path. My change does not load the 16x16 pixel image but the 72x72, which IMO should then be scaled to the desired size in global.css, so it can be overridden in a theme stylesheet and not be stuck with a tiny image that is blurry if you make it bigger.

Despite my opinion that this feature either needs to be removed or the file manager finished so you can at least organize/delete files, and until that happens people should be discouraged from using it all, at the very least it should display the file type icons.