Removed link loading global.css for each notification
global.css is _already_ loaded in the page head. Loading it over and over again with each notification item overrrides theme styling on the /notifications/system page, which messes up the styling of the Notifications drop-down under the bell icon too (but only on this page). Affects ALL themes because currently *none* of them include an override for this template.